fix: refactor steps to TestStepForm Componenet
