+++ frontend/src/components/admin/testManagementConfigMenu/customComponents/TestFormData.js
@@ -1,0 +1,38 @@
+export const TestFormData = {
+  testNameEnglish: "",
+  testNameFrench: "",
+  testReportNameEnglish: "",
+  testReportNameFrench: "",
+  testSection: "",
+  panels: [],
+  uom: "",
+  loinc: "",
+  resultType: "",
+  orderable: "Y",
+  notifyResults: "N",
+  inLabOnly: "N",
+  antimicrobialResistance: "N",
+  active: "Y",
+  dictionary: [],
+  dictionaryReference: "",
+  defaultTestResult: "",
+  sampleTypes: [],
+  lowValid: "-Infinity",
+  highValid: "Infinity",
+  lowReportingRange: "-Infinity",
+  highReportingRange: "Infinity",
+  lowCritical: "-Infinity",
+  highCritical: "Infinity",
+  significantDigits: "0",
+  resultLimits: [
+    {
+      ageRange: "0",
+      highAgeRange: "0",
+      gender: false,
+      lowNormal: "-Infinity",
+      highNormal: "Infinity",
+      lowNormalFemale: "-Infinity",
+      highNormalFemale: "Infinity",
+    },
+  ],
+};
